requirements.txt
flet==0.22.0
faster-whisper==1.0.3
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.4
pyperclip==1.8.2
pynput==1.7.7
pystray==0.19.5
Pillow==10.3.0
